requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
python-multipart==0.0.20
minio==7.2.14
SQLAlchemy==2.0.37
pyjwt==2.10.1
firebase-admin==6.6.0